backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
pypdf2>=3.0.1
groq>=0.4.2
python-multipart>=0.0.7
python-dotenv>=1.0.1
pydantic-settings>=2.1.0
mysql-connector-python>=8.3.0
sqlalchemy>=2.0.25
passlib[argon2]>=1.7.4
python-jose[cryptography]>=3.3.0
python-jwt>=2.0.1
